% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{coordinate_to_shapefile}
\alias{coordinate_to_shapefile}
\title{coordinate_to_shapefile}
\usage{
coordinate_to_shapefile(lat, long, country, ISO_level)
}
\description{
Convert a coordinate to the shapefile containing it.
}
