% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{create_location_file}
\alias{create_location_file}
\title{create_location_file}
\usage{
create_location_file(location, taxonomy_dir = "taxonomy-verified",
  shapefile, start_date, end_date, replace.existing = FALSE,
  verbose = TRUE)
}
\arguments{
\item{location}{Name of a location as a character}

\item{taxonomy_dir}{Name of a taxonomy directory (shapefiles will be stored here)}

\item{shapefile}{shapefile for location}
}
\description{
Create a location description file in the taxonomy for the filename
}
