% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{create_population_file}
\alias{create_population_file}
\title{create_population_file}
\usage{
create_population_file(data, taxonomy_dir = "taxonomy-verified",
  replace.existing = FALSE)
}
\arguments{
\item{taxonomy_dir}{Name of a taxonomy directory (shapefiles will be stored here)}

\item{location}{Name of a location as a character}
}
\description{
Create a population file in the taxonomy for the filename
}
