% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{create_worldpop_region_shapefiles}
\alias{create_worldpop_region_shapefiles}
\title{create_worldpop_region_shapefiles}
\usage{
create_worldpop_region_shapefiles()
}
\arguments{
\item{raster}{A Raster* object used to determine the extent and resolution}

\item{layers_dir}{Layers directory of the taxonomy folder}
}
\value{
A RasterLayer object of the same resolution and extent as \code{raster_layer}
}
\description{
Find distance to water estimates for a year and overlay them on a raster.
}
