% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{exists_shape_file}
\alias{exists_shape_file}
\title{exists_shape_file}
\usage{
exists_shape_file(location, taxonomy_dir = "taxonomy-verified",
  verbose = FALSE, date = now())
}
\arguments{
\item{location}{Name of a location as a character}

\item{taxonomy_dir}{Name of a taxonomy directory (shapefiles will be stored here)}

\item{date}{which date to look for a custom shapefile covering.}
}
\value{
boolean TRUE if a custom shapefile exists, FALSE otherwise
}
\description{
Check to see if a custom shapefile exists in the taxonomy already.
}
