% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{extract_shapefile_from_shapefile}
\alias{extract_shapefile_from_shapefile}
\title{extract_shapefile_from_shapefile}
\usage{
extract_shapefile_from_shapefile(shp1, shp2, strict = FALSE)
}
\arguments{
\item{shp1}{A shapefile to extract things from.}

\item{shp2}{The shapefile to use as a bounding box to determine what to extract from shp1}

\item{strict.}{boolean  If TRUE, take the intersection of the relevent parts of shp1 with shp2.  If FALSE, keep shapes from shp1 as they are, and include if any of the interior overlaps with the interior of shp2.}
}
\description{
Extract the parts of a shapefile that intersect another shapefile.
}
