% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{fix_location_name}
\alias{fix_location_name}
\title{fix_location_name}
\usage{
fix_location_name(location, taxonomy_dir = "taxonomy-verified",
  verbose = FALSE)
}
\arguments{
\item{location}{A character vector of names of a locations to fix}

\item{taxonomy_dir}{Name of a taxonomy directory (shapefiles will be stored here)}
}
\value{
A character containing the corrected filename or NA in case of failure
}
\description{
Attempt to look up a standardized representation of a location.
}
