% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.country.name.ISO2}
\alias{get.country.name.ISO2}
\title{get.country.name.ISO2}
\usage{
get.country.name.ISO2(ISO2)
}
\arguments{
\item{ISO2}{vector of ISO2s}
}
\value{
vector of country names matching ISO2. These can then be used to identify other characteristics of the country
}
\description{
Standard country name from ISO2.
}
