% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.country.name.ISO3}
\alias{get.country.name.ISO3}
\title{get.country.name.ISO3}
\usage{
get.country.name.ISO3(ISO)
}
\arguments{
\item{country}{vector of ISO3s}
}
\value{
vector of country names matching each ISO. These can then be used to identify other characteristics of the country
}
\description{
Standard country name from ISO3.
}
