% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.iso}
\alias{get.iso}
\title{get.iso}
\usage{
get.iso(country)
}
\arguments{
\item{country}{vector of country names}
}
\value{
vector of ISOs for each country. These can then be used to identify other characteristics of the country
}
\description{
get the official ISO for a vector of countries. This will identify most likely matched 
to the inputted country names used a variety of tools and methods.
}
