% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.ISO2.from.ISO3}
\alias{get.ISO2.from.ISO3}
\title{get.ISO2.from.ISO3}
\usage{
get.ISO2.from.ISO3(ISO3)
}
\arguments{
\item{ISO3}{vector of ISO3s}
}
\value{
vector of ISO2s matching each ISO3.
}
\description{
ISO2 from ISO3.
}
