% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.ISO3.from.ISO2}
\alias{get.ISO3.from.ISO2}
\title{get.ISO3.from.ISO2}
\usage{
get.ISO3.from.ISO2(ISO2)
}
\arguments{
\item{ISO2}{vector of ISO2s}
}
\value{
vector of ISO3s matching each ISO2 These can then be used to identify other characteristics of the country
}
\description{
ISO3 from ISO2.
}
