% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.location.std}
\alias{get.location.std}
\alias{get.city.std}
\title{get.location.std}
\usage{
get.location.std(location)

get.city.std(location)
}
\arguments{
\item{location}{vector of location names}

\item{location}{vector of city names}
}
\value{
vector of names for each location. These can then be used to identify other characteristics of the location

vector of names for each location. These can then be used to identify other characteristics of the location
}
\description{
get a standardized name for a vector of locations. This will identify most likely matched 
to the inputted names used a variety of tools and methods. This should be used when the data likely have
locations that are not only countries (i.e., some sub-countries like England or Dubai)

get a standardized name for a vector of Cities. This will identify most likely matched 
to the inputted names used a variety of tools and methods.
}
