% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get.region}
\alias{get.region}
\title{get.region}
\usage{
get.region(ISO)
}
\arguments{
\item{ISO}{vector of ISO3 codes}
}
\value{
vector of regions
}
\description{
Identify region from ISO
}
