% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get.subregion}
\alias{get.subregion}
\title{get.subregion}
\usage{
get.subregion(ISO)
}
\arguments{
\item{ISO}{vector of ISO3 codes}
}
\value{
vector of subregions
}
\description{
Identify subregion from ISO
}
