% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get.UNcode}
\alias{get.UNcode}
\title{get.UNcode}
\usage{
get.UNcode(country)
}
\arguments{
\item{country}{vector of country names}
}
\value{
vector of UN codes matching each country.
}
\description{
UN code from country
}
