% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{get_shapefile_cover}
\alias{get_shapefile_cover}
\title{get_shapefile_cover}
\usage{
get_shapefile_cover(country_names, taxonomy_dir = "taxonomy-verified",
  verbose = TRUE)
}
\arguments{
\item{country_names}{A character vector of country names to pull shapefiles for.}

\item{taxonomy_dir}{The path to the taxonomy filesystem to store the shapefile in.}

\item{verbose}{Whether to pass along messages about altered location names}
}
\description{
Get a shapefile covering the union of all of the countries in question.
}
