% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{load_covariate_estimates}
\alias{load_covariate_estimates}
\title{load_covariate_estimates}
\usage{
load_covariate_estimates(raster, region, layers_dir = "Layers",
  tol = 1e-05)
}
\arguments{
\item{raster}{A Raster* object used to determine the extent and resolution}

\item{layers_dir}{Layers directory of the taxonomy folder}
}
\value{
A RasterLayer object of the same resolution and extent as \code{raster_layer}
}
\description{
Find distance to water estimates for a year and overlay them on a raster.
}
