% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{load_population_estimates}
\alias{load_population_estimates}
\title{load_population_estimates}
\usage{
load_population_estimates(region, raster, year, layers_dir = "Layers")
}
\arguments{
\item{region}{The region for which underlying raster to use. one of "AFR" "ASI" or "AMR"}

\item{raster}{A Raster* object used to determine the extent and resolution}

\item{year}{Which year to try to estimate}

\item{layers_dir}{Layers directory of the taxonomy folder}
}
\value{
A RasterLayer object of the same resolution and extent as \code{raster}
}
\description{
Find population estimates for a year and overlay them on a raster.  This will take much longer on rasters without a calculable resolution
}
