% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{load_population_estimates_slow}
\alias{load_population_estimates_slow}
\title{load_population_estimates_slow}
\usage{
load_population_estimates_slow(raster_layer, year, layers_dir = "Layers")
}
\arguments{
\item{year}{Which year to try to estimate}

\item{layers_dir}{Layers directory of the taxonomy folder}

\item{raster}{A Raster* object used to determine the extent and resolution}
}
\value{
A RasterLayer object of the same resolution and extent as \code{raster_layer}
}
\description{
Find population estimates for a year and overlay them on a raster.  This will work on arbitrary rasters but is memory intensive
}
