% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{lookup_WHO_region}
\alias{lookup_WHO_region}
\title{lookup_WHO_region}
\usage{
lookup_WHO_region(x)
}
\arguments{
\item{x}{A vector of country or subcountry codes to look up the WHO region for}
}
\description{
lookup_WHO_region
}
