% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{match_country}
\alias{match_country}
\title{match_country}
\usage{
match_country(a, country_names = country_names, return_ISO = TRUE,
  return_country_name = FALSE, return_score = FALSE,
  return_score_matrix = FALSE)
}
\arguments{
\item{a}{country name to match}

\item{country_names}{country_names df loaded with the package}

\item{return_ISO}{TRUE/FALSE}

\item{return_country_name}{TRUE/FALSE}

\item{return_score}{TRUE/FALSE}

\item{return_score_matrix}{TRUE/FALSE}
}
\value{
ISOs, country names, matching scores, full matching distance matrix
}
\description{
use stringdist to get best match for country name
}
