% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{match_locs_level2}
\alias{match_locs_level2}
\title{match_locs_level2}
\usage{
match_locs_level2(a, names = locations_lvl2, return_ISO3 = FALSE,
  return_name = TRUE, return_Code = FALSE, return_score = FALSE,
  return_score_matrix = FALSE)
}
\arguments{
\item{a}{location name to match}

\item{names}{locations_lvl2 df loaded with the package}

\item{return_ISO3}{TRUE/FALSE}

\item{return_name}{TRUE/FALSE return standardized name}

\item{return_Code}{TRUE/FALSE}

\item{return_score}{TRUE/FALSE}

\item{return_score_matrix}{TRUE/FALSE}
}
\value{
ISOs, country names, matching scores, full matching distance matrix
}
\description{
use stringdist to get best match for country name, if not official country
}
